add Cart and CartItem

+++ store/src/main/java/app/store/repository/CartItemRepository.java
@@ -3,5 +3,5 @@ package app.store.repository;
 import app.store.entity.CartItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CartItemRepository extends JpaRepository<CartItem, Integer> {
+public interface CartItemRepository extends JpaRepository<CartItem, Long> {
 }
